Implement auth check in restaurants features

+++ middlewares/restaurants-handler.js
@@ -10,6 +10,7 @@ restaurantHandler.getAll = (req, res, next) => {
   const limit = 9;
   const sort = req.query.sort;
   const currentPage = parseInt(req.query.page) || 1;
+  const userId = req.user.id
   
   // 取得排序條件
   let condition = []
@@ -40,7 +41,7 @@ restaurantHandler.getAll = (req, res, next) => {
       {location: { [Op.like]: `%${keywords}%` }},
       {description: { [Op.like]: `%${keywords}%` }},
     ]
-  } : {}
+  } : { userId }
 
   return restaurantList.findAndCountAll({
     attributes: ['id', 'name', 'name_en', 'category','location', 'rating', 'image', 'description'],
@@ -78,11 +79,21 @@ restaurantHandler.getAll = (req, res, next) => {
 restaurantHandler.getById = async (req, res, next) => {
   try {
     const id = req.params.id;
-    req.restaurant = await restaurantList.findByPk(id, {
-    attributes: ['id', 'name', 'category', 'location','phone', 'google_map', 'description', 'image'],
-    raw:true});
-    next();
+    const userId = req.user.id
 
+    req.restaurant = await restaurantList.findByPk(id, {
+      attributes: ['id', 'name', 'category', 'location','phone', 'google_map', 'description', 'image', 'userId'],
+      raw:true});
+    
+    if (!req.restaurant) {
+      req.flash('error', '查無資料')
+      return res.redirect('/restaurants')
+    }
+    if (userId !== req.restaurant.userId) {
+      req.flash('error', '無檢視權限')
+      return res.redirect('/restaurants')
+    }
+    next();
   } catch (error) {
     error.errorMessage = '資料讀取失敗';
     next(error);
@@ -92,7 +103,19 @@ restaurantHandler.getById = async (req, res, next) => {
 restaurantHandler.editById = async (req, res, next) => {
   try {
     const id = req.params.id;
+    const userId = req.user.id;
+
     req.restaurant = await restaurantList.findByPk(id, {raw: true});
+
+    if (!req.restaurant) {
+      req.flash('error', '查無資料')
+      return res.redirect('/restaurants')
+    }
+    if (userId !== req.restaurant.userId) {
+      req.flash('error', '無編輯權限')
+      return res.redirect('/restaurants')
+    }
+
     next();
 
   } catch (error) {
@@ -105,13 +128,24 @@ restaurantHandler.update = async(req, res, next) => {
   try {
     const id = req.params.id;
     const {name, name_en, category, phone, image, location, google_map, rating, description} = req.body;
+    const userId = req.user.id
+    const restaurant = await restaurantList.findByPk(id, { attributes: [ 'id', 'userId' ]}) 
+    
+    if (!restaurant) {
+      req.flash('error', '查無資料')
+      return res.redirect('/restaurants')
+    }
+    if (userId !== restaurant.userId) {
+      req.flash('error', '無編輯權限')
+      return res.redirect('/restaurants')
+    }
     await restaurantList.update(
       {name, name_en, category, phone, image, location, google_map, rating, description}, {where: {id}});
     req.id = id;
     next();
 
   } catch (error) {
-    error.errorMessage = '更新失敗';
+    error.errorMessage = '編輯失敗';
     next(error);
   }
 }
@@ -119,7 +153,8 @@ restaurantHandler.update = async(req, res, next) => {
 restaurantHandler.create = async (req, res, next) => {
   try {
     const {name, name_en, category, phone, image, location, google_map, rating, description} = req.body;
-    await restaurantList.create({name, name_en, category, phone, image, location, google_map, rating, description});
+    const userId = req.user.id
+    await restaurantList.create({name, name_en, category, phone, image, location, google_map, rating, description, userId});
     next();
   } catch (error) {
     error.errorMessage = '新增失敗';
@@ -130,7 +165,17 @@ restaurantHandler.create = async (req, res, next) => {
 restaurantHandler.delete = async (req, res, next) => {
   try {
     const id = req.params.id;
-    console.log('id:', id)
+    const userId = req.user.id;
+    const restaurant = await restaurantList.findByPk(id, {attributes: ['id', 'userId']})
+    if (!restaurant) {
+      req.flash('error', '查無資料')
+      return res.redirect('/restaurants')
+    }
+    if (userId !== restaurant.userId) {
+      req.flash('error', '無刪除權限')
+      return res.redirect('/restaurants')
+    }
+
     await restaurantList.destroy({where: {id}});
     next();
   } catch (error) {
